add after_remove property in config

+++ runner/engine.go
@@ -314,8 +314,11 @@ func (e *Engine) runBin() error {
 		if _, err = os.Stat(cmdBinPath); os.IsNotExist(err) {
 			return
 		}
-		if err = os.Remove(cmdBinPath); err != nil {
-			e.mainLog("failed to remove %s, error: %s", e.config.rel(e.config.binPath()), err)
+
+		if e.config.Build.Remove {
+			if err = os.Remove(cmdBinPath); err != nil {
+				e.mainLog("failed to remove %s, error: %s", e.config.rel(e.config.binPath()), err)
+			}
 		}
 	}(cmd)
 	return nil
